+++ bee-tangle/src/lib.rs
@@ -5,8 +5,10 @@
 
 #![deny(missing_docs)]
 
+mod message_data;
 mod tangle;
 mod walker;
 
+pub use message_data::MessageData;
 pub use tangle::Tangle;
 pub use walker::TangleWalker;
