+++ bee-tangle/tests/tangle.rs
@@ -12,8 +12,8 @@ fn insert_get() {
     let tangle = Tangle::new();
     tangle.insert(message_id, message.clone(), metadata.clone());
 
-    let (fetched_message, fetched_metadata) = tangle.get(&message_id).unwrap();
+    let message_data = tangle.get(&message_id).unwrap();
 
-    assert_eq!(*fetched_message, message);
-    assert_eq!(fetched_metadata, metadata);
+    assert_eq!(message_data.message(), &message);
+    assert_eq!(message_data.metadata(), &metadata);
 }
